refactor: adjusted the structure
