+++ vite.config.ts
@@ -1,12 +1,10 @@
 /// <reference types="vitest" />
 
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
-  test:{
-
-  }
-})
+  test: {},
+});
